Move ParseExec to the client where it is used.
